+++ lib/chef/handler/datadog.rb
@@ -79,6 +79,18 @@ class Chef
       # @param run_status [Chef::RunStatus] current run status
       # @return [Array] alert_type, event_priority, event_title, event_body
       def build_event_data(hostname, run_status)
+        # bail early in case of a compiletime failure
+        # OPTIMIZE: Use better inspectors to handle failure scenarios, refactor needed.
+        if run_status.elapsed_time.nil?
+
+          alert_type = 'error'
+          event_title = "Chef failed during compile phase on #{hostname} "
+          event_priority = 'normal'
+          event_body = 'Chef was unable to complete a run, an error during compilation may have occured.'
+
+          return [alert_type, event_priority, event_title, event_body]
+        end
+
         run_time = pluralize(run_status.elapsed_time, 'second')
 
         # This is the first line of the Event body, the rest is appended here.
@@ -155,6 +167,11 @@ class Chef
       # @param hostname [String] resolved hostname to attach to series
       # @param run_status [Chef::RunStatus] current run status
       def emit_metrics_to_datadog(hostname, run_status)
+        # If there is a failure during compile phase, a large portion of
+        # run_status may be unavailable. Bail out here
+        warn_msg = 'Error during compile phase, no Datadog metrics available.'
+        return Chef::Log.warn(warn_msg) if run_status.elapsed_time.nil?
+
         @dog.emit_point('chef.resources.total', run_status.all_resources.length, :host => hostname)
         @dog.emit_point('chef.resources.updated', run_status.updated_resources.length, :host => hostname)
         @dog.emit_point('chef.resources.elapsed_time', run_status.elapsed_time, :host => hostname)