+++ spec/datadog_spec.rb
@@ -331,6 +331,36 @@ describe Chef::Handler::Datadog, :vcr => :new_episodes do
     end
   end
 
+  describe 'resources' do
+    before(:each) do
+      @node = Chef::Node.build('chef.handler.datadog.test-resources')
+      @node.send(:chef_environment, 'resources')
+      @events = Chef::EventDispatch::Dispatcher.new
+      @run_context = Chef::RunContext.new(@node, {}, @events)
+      @run_status = Chef::RunStatus.new(@node, @events)
+    end
+
+    context 'failure during compile phase' do
+      before(:each) do
+        @handler.run_report_unsafe(@run_status)
+      end
+
+      it 'does not emit metrics' do
+        expect(a_request(:post, METRICS_ENDPOINT).with(
+          :query => { 'api_key' => @handler.config[:api_key] }
+        )).to_not have_been_made
+      end
+
+      it 'posts an event' do
+        expect(a_request(:post, EVENTS_ENDPOINT).with(
+          :query => { 'api_key' => @handler.config[:api_key] },
+          :body => hash_including(:msg_text => 'Chef was unable to complete a run, an error during compilation may have occured.'),
+          :body => hash_including(:msg_title => "Chef failed during compile phase on #{@node.name} "),
+        )).to have_been_made.times(1)
+      end
+    end
+  end
+
     # TODO: test failures:
     # @run_status.exception = Exception.new('Boy howdy!')
 end